# crates/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Silicon Apps API",
    "version": "2026-10-09",
    "summary": "Create, publish, find, install and follow command-line apps for Carbons and Silicons.",
...
    "contact": {
      "name": "Team of Silicons",
      "url": "https://teamofsilicons.com",
      "email": "bugs@teamofsilicons.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/teamofsilicons/silicon-apps/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://apps.teamofsilicons.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    },
    {}
  ],
...
  "paths": {
...
    "/v1/apps/{app_id}/events": {
      "get": {
        "operationId": "listAppEvents",
        "summary": "List an app's events",
        "description": "The app's event log as JSON pages: releases, package validation steps with the three commands' results, author invitations, joins and leaves, access changes and more. Authors only.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppId"
          },
          {
            "$ref": "#/components/parameters/After"
          },
          {
            "$ref": "#/components/parameters/Types"
          },
          {
            "$ref": "#/components/parameters/EventLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/apps/{app_id}/events/stream": {
      "get": {
        "operationId": "streamAppEvents",
        "summary": "Stream an app's events",
        "description": "Live events for an app's authors. Server-sent events. Each event has `id` (its seq), `event` (its type) and `data` (the Event as JSON). The stream starts with a `: ready cursor=N` comment and sends a `: heartbeat` comment every 15 seconds. Without Last-Event-ID it starts at the newest event. It closes after 30 minutes; reconnect with Last-Event-ID to continue.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AppId"
          },
          {
            "$ref": "#/components/parameters/Types"
          },
          {
            "$ref": "#/components/parameters/LastEventIdHeader"
          },
          {
            "$ref": "#/components/parameters/LastEventIdQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "An event stream.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "id: 42\nevent: release.promoted\ndata: {\"seq\":42,\"type\":\"release.promoted\",\"app_id\":\"briefcase\",\"data\":{\"channel\":\"production\",\"version\":\"1.4.0\"}}\n\n"
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/events": {
      "get": {
        "operationId": "listMyEvents",
        "summary": "List your account's events",
        "description": "Your account feed as JSON pages: invitations to you, everything about apps you author, and public events (releases, publishing) of apps you installed. With `subscription`, that subscription's events.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/After"
          },
          {
            "$ref": "#/components/parameters/Types"
          },
          {
            "$ref": "#/components/parameters/EventLimit"
          },
          {
            "name": "subscription",
            "in": "query",
            "required": false,
            "description": "Read through one of your subscriptions' filters.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/events/stream": {
      "get": {
        "operationId": "streamMyEvents",
        "summary": "Stream your account's events",
        "description": "Live events for the signed-in account. With `subscription`, the subscription's events; without Last-Event-ID it resumes where that subscription's stream last stopped. Server-sent events. Each event has `id` (its seq), `event` (its type) and `data` (the Event as JSON). The stream starts with a `: ready cursor=N` comment and sends a `: heartbeat` comment every 15 seconds. Without Last-Event-ID it starts at the newest event. It closes after 30 minutes; reconnect with Last-Event-ID to continue.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Types"
          },
          {
            "$ref": "#/components/parameters/LastEventIdHeader"
          },
          {
            "$ref": "#/components/parameters/LastEventIdQuery"
          },
          {
            "name": "subscription",
            "in": "query",
            "required": false,
            "description": "Stream one of your subscriptions.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An event stream.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "id: 42\nevent: release.promoted\ndata: {\"seq\":42,\"type\":\"release.promoted\",\"app_id\":\"briefcase\",\"data\":{\"channel\":\"production\",\"version\":\"1.4.0\"}}\n\n"
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimitLimit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimitRemaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimitReset"
              },
              "Apps-Version": {
                "$ref": "#/components/headers/AppsVersion"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Silicon Accounts access token for the silicon-apps audience."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apps_session",
        "description": "The store's HttpOnly browser session. Browser mutations also need an allowed Origin."
      }
    },
    "schemas": {
...
      "EventPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "next_after": {
            "type": "integer",
            "description": "Pass as `after` for the next page."
          },
          "has_more": {
            "type": "boolean"
          },
          "cursor": {
            "type": "integer",
            "description": "The newest event seq in the log."
          }
        },
        "required": [
          "items",
          "next_after",
          "has_more",
          "cursor"
        ]
      },
...
    },
    "parameters": {
      "AppId": {
        "name": "app_id",
        "in": "path",
        "required": true,
        "description": "The app's permanent ID.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 64
        }
      },
...
      "Types": {
        "name": "types",
        "in": "query",
        "required": false,
        "description": "Comma-separated event types to include: exact types, `group.*` (for example `release.*`) or `*`.",
        "schema": {
          "type": "string"
        },
        "example": "release.promoted,package.*"
      },
...
      "After": {
        "name": "after",
        "in": "query",
        "required": false,
        "description": "Return events after this seq. Defaults to 0, the oldest.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      },
      "EventLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Events per page, 1 to 500.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      }
    },
...
  }
}